Write tests for process file and process url
